+++ mex/editor/rules/transform.py
@@ -38,14 +38,14 @@ from mex.common.types import (
     Text,
     TextLanguage,
 )
-from mex.editor.edit.models import (
+from mex.editor.models import MODEL_CONFIG_BY_STEM_TYPE, EditorValue
+from mex.editor.rules.models import (
     EditorField,
     EditorPrimarySource,
     InputConfig,
     ValidationMessage,
 )
-from mex.editor.models import EditorValue
-from mex.editor.transform import ensure_list, transform_value
+from mex.editor.transform import ensure_list, transform_value, transform_values
 from mex.editor.types import AnyModelValue
 
 
@@ -223,7 +223,7 @@ def transform_models_to_fields(
     mergeable_fields = sorted(
         {
             f
-            for e in extracted_items
+            for e in [*extracted_items, additive]
             for f in MERGEABLE_FIELDS_BY_CLASS_NAME[e.entityType]
         }
     )
@@ -317,7 +317,7 @@ def _transform_editor_value_to_model_value(
             if vocabulary_name := (value.badge or input_config.badge_default):
                 return cast("type[AnyVocabularyEnum]", vocabulary)[vocabulary_name]
     if field_name in TEMPORAL_FIELDS_BY_CLASS_NAME[class_name]:
-        precision = TemporalEntityPrecision(value.badge)
+        precision = TemporalEntityPrecision(value.badge or input_config.badge_default)
         temporal_class = TEMPORAL_ENTITY_CLASSES_BY_PRECISION[precision]
         return temporal_class(str(value.text), precision=precision)
     if field_name in REFERENCE_FIELDS_BY_CLASS_NAME[class_name]:
@@ -389,3 +389,23 @@ def transform_validation_error_to_messages(
         )
         for error in error.errors()
     ]
+
+
+def transform_fields_to_title(
+    stem_type: str,
+    fields: list[EditorField],
+) -> list[EditorValue]:
+    """Convert a list of editor fields into title values based on the title config."""
+    config = MODEL_CONFIG_BY_STEM_TYPE[stem_type]
+    titles: list[EditorValue] = []
+    for field in fields:
+        if field.name == config.title:
+            titles = [
+                value
+                for primary_source in field.primary_sources
+                for value in primary_source.editor_values
+            ]
+            break
+    if titles:
+        return titles
+    return transform_values(stem_type)