+++ mex/editor/state.py
@@ -21,6 +21,11 @@ class State(rx.State):
             raw_path="/?page=1",
         ),
         NavItem(
+            title="Create",
+            path="/create",
+            raw_path="/create/",
+        ),
+        NavItem(
             title="Edit",
             path="/item/[identifier]",
             raw_path=f"/item/{MEX_PRIMARY_SOURCE_STABLE_TARGET_ID}/",
@@ -31,9 +36,14 @@ class State(rx.State):
             raw_path="/merge_items/",
         ),
         NavItem(
-            title="Aux Search",
-            path="/aux-search",
-            raw_path="/aux-search/",
+            title="Merge Items",
+            path="/merge_items",
+            raw_path="/merge_items/",
+        ),
+        NavItem(
+            title="Ingest",
+            path="/ingest",
+            raw_path="/ingest/",
         ),
     ]