+++ CHANGELOG.md
@@ -11,13 +11,32 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- rename aux-search component to ingest, matching naming in backend
+
 ### Deprecated
 
 ### Removed
 
 ### Fixed
 
+- fix transparency of overlay containers
+
 ### Security
+
+## [0.18.0] - 2025-06-17
+
+### Changes
+
+- install mex-common and mex-artificial from pypi
+
+## [0.17.0] - 2025-06-16
+
+### Added
+
+- import artificial data helper and use in test
+- page to create new rule-set in editor
+- new rules module containing shared code of create and edit pages
+- button for switching between rendered and input view for additive field values
 
 ## [0.16.0] - 2025-05-22